Refactored codebase with OZ's Prettier config

+++ contracts/CentralizedAppealableArbitrator.sol
@@ -23,7 +23,7 @@ interface IArbitrable {
      *  @param _metaEvidenceID Unique identifier of meta-evidence.
      *  @param _evidence A link to the meta-evidence JSON.
      */
-    event MetaEvidence(uint indexed _metaEvidenceID, string _evidence);
+    event MetaEvidence(uint256 indexed _metaEvidenceID, string _evidence);
 
     /** @dev To be emmited when a dispute is created to link the correct meta-evidence to the disputeID
      *  @param _arbitrator The arbitrator of the contract.
@@ -33,9 +33,9 @@ interface IArbitrable {
      */
     event Dispute(
         Arbitrator indexed _arbitrator,
-        uint indexed _disputeID,
-        uint _metaEvidenceID,
-        uint _evidenceGroupID
+        uint256 indexed _disputeID,
+        uint256 _metaEvidenceID,
+        uint256 _evidenceGroupID
     );
 
     /** @dev To be raised when evidence are submitted. Should point to the ressource (evidences are not to be stored on chain due to gas considerations).
@@ -46,7 +46,7 @@ interface IArbitrable {
      */
     event Evidence(
         Arbitrator indexed _arbitrator,
-        uint indexed _evidenceGroupID,
+        uint256 indexed _evidenceGroupID,
         address indexed _party,
         string _evidence
     );
@@ -56,18 +56,14 @@ interface IArbitrable {
      *  @param _disputeID ID of the dispute in the Arbitrator contract.
      *  @param _ruling The ruling which was given.
      */
-    event Ruling(
-        Arbitrator indexed _arbitrator,
-        uint indexed _disputeID,
-        uint _ruling
-    );
+    event Ruling(Arbitrator indexed _arbitrator, uint256 indexed _disputeID, uint256 _ruling);
 
     /** @dev Give a ruling for a dispute. Must be called by the arbitrator.
      *  The purpose of this function is to ensure that the address calling it has the right to rule on the contract.
      *  @param _disputeID ID of the dispute in the Arbitrator contract.
      *  @param _ruling Ruling given by the arbitrator. Note that 0 is reserved for "Not able/wanting to make a decision".
      */
-    function rule(uint _disputeID, uint _ruling) external;
+    function rule(uint256 _disputeID, uint256 _ruling) external;
 }
 
 /** @title Arbitrator
@@ -85,18 +81,12 @@ abstract contract Arbitrator {
     }
 
     modifier requireArbitrationFee(bytes calldata _extraData) {
-        require(
-            msg.value >= this.arbitrationCost(_extraData),
-            "Not enough ETH to cover arbitration costs."
-        );
+        require(msg.value >= this.arbitrationCost(_extraData), "Not enough ETH to cover arbitration costs.");
         _;
     }
 
-    modifier requireAppealFee(uint _disputeID, bytes calldata _extraData) {
-        require(
-            msg.value >= this.appealCost(_disputeID, _extraData),
-            "Not enough ETH to cover appeal costs."
-        );
+    modifier requireAppealFee(uint256 _disputeID, bytes calldata _extraData) {
+        require(msg.value >= this.appealCost(_disputeID, _extraData), "Not enough ETH to cover appeal costs.");
         _;
     }
 
@@ -104,27 +94,18 @@ abstract contract Arbitrator {
      *  @param _disputeID ID of the dispute.
      *  @param _arbitrable The contract which created the dispute.
      */
-    event DisputeCreation(
-        uint indexed _disputeID,
-        IArbitrable indexed _arbitrable
-    );
+    event DisputeCreation(uint256 indexed _disputeID, IArbitrable indexed _arbitrable);
 
     /** @dev To be raised when a dispute can be appealed.
      *  @param _disputeID ID of the dispute.
      */
-    event AppealPossible(
-        uint indexed _disputeID,
-        IArbitrable indexed _arbitrable
-    );
+    event AppealPossible(uint256 indexed _disputeID, IArbitrable indexed _arbitrable);
 
     /** @dev To be raised when the current ruling is appealed.
      *  @param _disputeID ID of the dispute.
      *  @param _arbitrable The contract which created the dispute.
      */
-    event AppealDecision(
-        uint indexed _disputeID,
-        IArbitrable indexed _arbitrable
-    );
+    event AppealDecision(uint256 indexed _disputeID, IArbitrable indexed _arbitrable);
 
     /** @dev Create a dispute. Must be called by the arbitrable contract.
      *  Must be paid at least arbitrationCost(_extraData).
@@ -132,29 +113,25 @@ abstract contract Arbitrator {
      *  @param _extraData Can be used to give additional info on the dispute to be created.
      *  @return disputeID ID of the dispute created.
      */
-    function createDispute(uint _choices, bytes calldata _extraData)
+    function createDispute(uint256 _choices, bytes calldata _extraData)
         public
         payable
         virtual
         requireArbitrationFee(_extraData)
-        returns (uint disputeID)
+        returns (uint256 disputeID)
     {}
 
     /** @dev Compute the cost of arbitration. It is recommended not to increase it often, as it can be highly time and gas consuming for the arbitrated contracts to cope with fee augmentation.
      *  @param _extraData Can be used to give additional info on the dispute to be created.
      *  @return fee Amount to be paid.
      */
-    function arbitrationCost(bytes calldata _extraData)
-        public
-        view
-        virtual
-        returns (uint fee);
+    function arbitrationCost(bytes calldata _extraData) public view virtual returns (uint256 fee);
 
     /** @dev Appeal a ruling. Note that it has to be called before the arbitrator contract calls rule.
      *  @param _disputeID ID of the dispute to be appealed.
      *  @param _extraData Can be used to give extra info on the appeal.
      */
-    function appeal(uint _disputeID, bytes calldata _extraData)
+    function appeal(uint256 _disputeID, bytes calldata _extraData)
         public
         payable
         virtual
@@ -168,43 +145,26 @@ abstract contract Arbitrator {
      *  @param _extraData Can be used to give additional info on the dispute to be created.
      *  @return fee Amount to be paid.
      */
-    function appealCost(uint _disputeID, bytes calldata _extraData)
-        public
-        view
-        virtual
-        returns (uint fee);
+    function appealCost(uint256 _disputeID, bytes calldata _extraData) public view virtual returns (uint256 fee);
 
     /** @dev Compute the start and end of the dispute's current or next appeal period, if possible.
      *  @param _disputeID ID of the dispute.
      *  @return start The start of the period.
      *  @return end The end of the period.
      */
-    function appealPeriod(uint _disputeID)
-        public
-        view
-        virtual
-        returns (uint start, uint end)
-    {}
+    function appealPeriod(uint256 _disputeID) public view virtual returns (uint256 start, uint256 end) {}
 
     /** @dev Return the status of a dispute.
      *  @param _disputeID ID of the dispute to rule.
      *  @return status The status of the dispute.
      */
-    function disputeStatus(uint _disputeID)
-        public
-        view
-        virtual
-        returns (DisputeStatus status);
+    function disputeStatus(uint256 _disputeID) public view virtual returns (DisputeStatus status);
 
     /** @dev Return the current ruling of a dispute. This is useful for parties to know if they should appeal.
      *  @param _disputeID ID of the dispute.
      *  @return ruling The ruling which has been given or the one which will be given if there is no appeal.
      */
-    function currentRuling(uint _disputeID)
-        public
-        view
-        virtual
-        returns (uint ruling);
+    function currentRuling(uint256 _disputeID) public view virtual returns (uint256 ruling);
 }
 
 /** @title Centralized Arbitrator
@@ -252,12 +212,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
      *  @param _extraData Not used by this contract.
      *  @return fee Amount to be paid.
      */
-    function arbitrationCost(bytes memory _extraData)
-        public
-        view
-        override
-        returns (uint fee)
-    {
+    function arbitrationCost(bytes memory _extraData) public view override returns (uint256 fee) {
         return arbitrationPrice;
     }
 
@@ -266,12 +221,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
      *  @param _extraData Not used by this contract.
      *  @return fee Amount to be paid.
      */
-    function appealCost(uint256 _disputeID, bytes memory _extraData)
-        public
-        view
-        override
-        returns (uint256 fee)
-    {
+    function appealCost(uint256 _disputeID, bytes memory _extraData) public view override returns (uint256 fee) {
         return arbitrationPrice;
     }
 
@@ -331,10 +281,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
         DisputeStruct storage dispute = disputes[_disputeID];
 
         require(_ruling <= dispute.choices, "Invalid ruling.");
-        require(
-            dispute.status != DisputeStatus.Solved,
-            "The dispute must not be solved already."
-        );
+        require(dispute.status != DisputeStatus.Solved, "The dispute must not be solved already.");
 
         dispute.ruling = _ruling;
 
@@ -347,8 +294,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
             dispute.status = DisputeStatus.Solved;
 
             payable(msg.sender).send(dispute.fee); // Avoid blocking.
-            if (dispute.isAppealed == true)
-                payable(msg.sender).send(dispute.fee); // Appeal fee
+            if (dispute.isAppealed == true) payable(msg.sender).send(dispute.fee); // Appeal fee
 
             dispute.arbitrated.rule(_disputeID, _ruling);
         }
@@ -358,12 +304,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
      *  @param _disputeID ID of the dispute to rule.
      *  @return status The status of the dispute.
      */
-    function disputeStatus(uint256 _disputeID)
-        public
-        view
-        override
-        returns (DisputeStatus status)
-    {
+    function disputeStatus(uint256 _disputeID) public view override returns (DisputeStatus status) {
         return disputes[_disputeID].status;
     }
 
@@ -371,12 +312,7 @@ contract CentralizedAppealableArbitrator is Arbitrator {
      *  @param _disputeID ID of the dispute to rule.
      *  @return ruling The ruling which would or has been given.
      */
-    function currentRuling(uint256 _disputeID)
-        public
-        view
-        override
-        returns (uint256 ruling)
-    {
+    function currentRuling(uint256 _disputeID) public view override returns (uint256 ruling) {
         return disputes[_disputeID].ruling;
     }
 }
